Make sure key rotation query filters by computer
Right now the key rotation query gets the latest secret filtering only by secret type and not also by computer, which means the latest date escrowed could be the latest for any computer. This leads to some clients getting the instruction to rotate "viewed" recovery keys that aren't actually viewed, and other clients not getting the instruction to rotate a viewed recovery key.

This PR adds in an additional filter to make sure the date secret fetched for `rotation_required` is for the actual machine making the request.